requirements.txt
aiohttp
azure-identity
azure-mgmt-resourcegraph
fastapi
httpx
uvicorn
pyyaml
pandas
numpy
plotly
streamlit
pytest
pytest-asyncio
